requirements.txt
numpy==1.21.0
pandas==1.3.0
biopython==1.79
